chore: isShipping..Completed inside ShippingForm
